misc: Ignore ENOTSUP when copying file metadata
This error will be thrown when, for example, trying to copy files to
an FTP mount. Getting this error is a very strong indication that
doing this metadata modification on this location will *never* work,
so there's not a lot of point in showing the user an error.

(cherry picked from commit 47071a49a02da51137bb4edd0dbec006668ac013)