Default Pango backend to fontconfig on Windows (#445, #453)
The Pango win32 font backend has some long-standing issues with fallback
fonts (#445) and also has a current problem with line heights for empty
lines (#453). While there have been problems in the past with fontconfig
on Windows, it looks like it's currently a better option for us.

See https://bugzilla.gnome.org/show_bug.cgi?id=785313 for an old bug
related to slowness with fontconfig. The fix there was to switch to the
native win32 fonts backend, but since that has issues, we're explicitly
switching the other direction for the moment.

This uses the environment variable method of switching backends rather
than manually setting the font map, partially for ease and partially
because it's something the user can then manually override.