icons: embed application icon as a resource too
Before this patch not-installed meld or meld built for windows
has problems loading its own svg icon,
since the icon doesn't present in resources
and were never installed into the OS.

This leads to lack of application icon on non-Gnome linux DE's
and lack of icon inside about window on all platforms.

This patch embeds application icon into resources,
so it is usable from inside meld even if it is not installed.

This change can lead to a situation when builtin icon
overrides the icon from a user-configured icon theme.