dirdiff: Move pane action bar to use MeldFileButton and PathLabel
The goal here is to make the folder comparisons action bar UI the same
as the UI for file comparisons. We're also looking to fix the same kinds
of problems that the file UI change fixes, namely: access to the full
path, better overall display, and avoiding the deprecated
GtkFileChooserButton.

This change also moves to having a GObject property for the folders
being compared. Currently we don't really make great use of this, but
it gives us a starting point for migrating to a more property-driven
model here.