build: Use dictionaries in configuration_data objects
Since 0.49.0[0], a dictionary could be used as a replacement for
each key/value pair as if `set` method was called for each of them.

All of the `configuration_data` objects have been changed to use a
dictionary, and those that with only one key, value pair have been
used directly.

[0] http://mesonbuild.com/Reference-manual.html#configuration_data