filediff, matchers.helpers: Improve matcher process clean up (#565)
We try to handle shut down and clean up of our multiprocessing matcher
process when we remove FileDiff tabs, but this wasn't working properly.
We know, for a start, that we have some garbage collection issues around
FileDiff in general.

Since __del__ is not guaranteed to get called (and in fact was not
getting called at all for this specific case in some minimal testing),
this commit moves the logic to a specific stop() call, which is invoked
in our FileDiff close handling.

This commit also cleans up our handling slightly to avoid spurious
errors when we close a file comparison tab before the matcher process
has been started (since it starts on idle).