Make filter headerbar buttons only visible in their related views
This is a very debatable change. I'm weighing up a desire for the
header bar to not be too wide (because it gives us a minimum size that
we can't easily work around) with the desire to have the UI be stable.
In this case, I've gone with having the available buttons change per
view, on the basis that there are three text buttons here and they take
up really quite a lot of horizontal allocation.

Ideally, I think this would be responsive and we'd try to always show
all three buttons, collapsing down to just the relevant one(s) if
necessary... but that's really quite a lot more work.