pathlabel: Format paths as home-relative where sensible
This changes the formatting of two cases, being when we're showing the
parent file path, and when we don't have a common parent set at all. In
reality, the first of these currently can't happen, because we only show
the parent when it's the root. The second is the interesting case here,
and only really affects version control view, making it so that viewing
version controlled folders under a user's home folder summarises the
path nicely.