vcview: Rework VC action bar buttons
This removes the delete-a-local-file action from our action bar, and
changes the resolve conflicts action to be an icon button in the same
linked block as other selection-based VC actions.

The rationale here is that deleting a file is a relatively uncommon
thing to do, and in any case isn't fundamental to the operation of Meld
as a VC viewer. It's also still available from the context menu, which
is how such actions are presented in e.g., Nautilus.

By contrast, resolving a conflict is very much a VC operation and while
it's comparatively rare... when it's needed it's very much needed.