dirdiff: Identify in-progress scans on tree parent row (#620)
This is a supplement to our toolbar spinner, since the latter is not
that obvious and also indicates ongoing tasks anywhere in the
application, rather than just the current pane.

It would possibly be nicer to have this presented as a spinner cell
renderer, but since we're only using this for one row (currently) that
seems like huge overkill in terms of additional cell data and renderers
for large trees. In the future if we move to more granular progress on
a per-file basis, per-cell spinners might make more sense.