dirdiff: Re-jig comparison option usage
The main change here is making comparison options not be an attribute,
but rather just recreate it when doing our scan. The reason here is that
it's nicer to not have duplicate state on the class, and it's somewhat
better to have the action state be explicit.

This change also fixes a linting error that was breaking CI.